requirements.txt
coreapi==2.3.3
coreschema==0.0.4
ddt==1.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.1
django-filter==2.1.0
django-oauth-toolkit==1.2.0
django-redis==4.10.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.2
flake8==3.7.8
gunicorn==19.9.0
psycopg2-binary==2.8.3
PyJWT==1.4.2
pytest==5.1.1
pytest-django==3.5.1
redis==3.3.8
social-auth-app-django==3.1.0
social-auth-core==3.2.0
whitenoise==4.1.2
coverage==4.5.4
pytest-cov==2.7.1